Добавлен вывод полезных полей. Вероятность ошибки и суммы бит

+++ checkdata.h
@@ -42,10 +42,10 @@ private:
     int countValidity_Ch1;
     int countValidity_Ch2;
     double validity_1;
-    double validityTrue_1;
+    int validityTrue_1;
     double validityAll_1;
     double validity_2;
-    double validityTrue_2;
+    int validityTrue_2;
     double validityAll_2;
